restructure and modernize cmake setup. Add workflow
makes c++ code warning free

+++ src/per/impl/per.cpp
@@ -1,5 +1,5 @@
 
-#include "per/per.h"
+#include "per/per.hpp"
 
 #include <utility>
 
@@ -31,7 +31,7 @@ void PrioritizedExperience::push(py::object value)
 {
    auto deleted_entry = m_sumtree.insert(
       value_type{
-         std::move(value), std::pow(m_max_priority / m_sumtree.total() * m_capacity, m_beta)},
+         std::move(value), std::pow(m_max_priority / m_sumtree.total() * static_cast<double>(m_capacity), m_beta)},
       m_max_priority);
 
    if(deleted_entry.has_value()) {
